+++ src/OpenTelemetry.Instrumentation.Owin/OwinInstrumentationOptions.cs
@@ -52,7 +52,7 @@ public class OwinInstrumentationOptions
     /// Gets or sets a value indicating whether the exception will be recorded as <see cref="ActivityEvent"/> or not.
     /// </summary>
     /// <remarks>
-    /// https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/exceptions.md.
+    /// https://github.com/open-telemetry/semantic-conventions/blob/main/docs/exceptions/exceptions-spans.md.
     /// </remarks>
     public bool RecordException { get; set; }