+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -10,6 +10,7 @@ Released 2025-Oct-29
   `service.name`, `service.instanceId`, and custom attributes.
   ([#3214](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3214))
   * Add WantedResourceAttributes to filter resource attributes to send to Geneva
+    ([#3367](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3367))
 * Allow custom string size limit in custom fields.
   ([#3360](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3360))