+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,15 +2,22 @@
 
 ## Unreleased
 
+## 1.13.1
+
+Released 2025-Oct-29
+
+* Support for specifying resource attributes, including
+  `service.name`, `service.instanceId`, and custom attributes.
+  ([#3214](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3214))
+* Allow custom string size limit in custom fields.
+  ([#3360](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3360))
+
 ## 1.13.0
 
 Released 2025-Oct-13
 
 * Updated OpenTelemetry core component version(s) to `1.13.1`.
   ([#3218](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3218))
-* Support for specifying resource attributes, including
-  `service.name`, `service.instanceId`, and custom attributes.
-  ([#3214](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3214))
 
 ## 1.13.0-alpha.1